refactor: move ignore related commands
