+++ services/interactions/src/interactions/config/index.ts
@@ -1,4 +1,5 @@
 export * from './banword';
+export * from './config-automod-ignores';
 export * from './config';
 export * from './filter';
 export * from './punishments';
